Split SacrificeScreen into new file
